Moved UserRepository to repository package
